test : Custom chart spec
