+++ app/client/cypress/integration/Smoke_TestSuite/ClientSideTests/DisplayWidgets/Chart_spec.js
@@ -319,75 +319,6 @@ describe("Chart Widget Functionality", function() {
       .should("exist");
   });
 
-  it("Custom Chart Widget Functionality", function() {
-    //changing the Chart type
-    cy.get(widgetsPage.toggleChartType).click({ force: true });
-    cy.UpdateChartType("Custom Chart");
-
-    cy.testJsontext(
-      "customfusionchartconfiguration",
-      `{{${JSON.stringify(this.data.ChartCustomConfig)}}}`,
-    );
-
-    //Verifying X-axis labels
-    cy.get(viewWidgetsPage.chartWidget).should("have.css", "opacity", "1");
-    const labels = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul"];
-    [0, 1, 2, 3, 4, 5, 6].forEach((k) => {
-      cy.get(viewWidgetsPage.rectangleChart)
-        .eq(k)
-        .trigger("mousemove", { force: true });
-      cy.get(viewWidgetsPage.Chartlabel)
-        .eq(k)
-        .should("have.text", labels[k]);
-    });
-    cy.PublishtheApp();
-  });
-
-  it("Toggle JS - Custom Chart Widget Functionality", function() {
-    //changing the Chart type
-    cy.UpdateChartType("Pie Chart");
-    cy.get(widgetsPage.toggleChartType).click({ force: true });
-    cy.testJsontext("charttype", "CUSTOM_FUSION_CHART");
-    cy.wait("@updateLayout").should(
-      "have.nested.property",
-      "response.body.responseMeta.status",
-      200,
-    );
-    cy.get(viewWidgetsPage.Chartlabel + ":first-child", {
-      timeout: 10000,
-    }).should("have.css", "opacity", "1");
-    //Verifying X-axis labels
-    cy.get(viewWidgetsPage.chartWidget).should("have.css", "opacity", "1");
-    const labels = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul"];
-    [0, 1, 2, 3, 4, 5, 6].forEach((k) => {
-      cy.get(viewWidgetsPage.rectangleChart)
-        .eq(k)
-        .trigger("mousemove", { force: true });
-      cy.get(viewWidgetsPage.Chartlabel)
-        .eq(k)
-        .should("have.text", labels[k]);
-    });
-
-    //Close edit prop
-    cy.get(commonlocators.editPropCrossButton).click();
-    cy.PublishtheApp();
-  });
-
-  it("Chart-Copy Verification", function() {
-    const modifierKey = Cypress.platform === "darwin" ? "meta" : "ctrl";
-    //Copy Chart and verify all properties
-    cy.copyWidget("chartwidget", viewWidgetsPage.chartWidget);
-
-    cy.PublishtheApp();
-  });
-
-  it("Chart-Delete Verification", function() {
-    // Delete the Chart widget
-    cy.deleteWidget(viewWidgetsPage.chartWidget);
-    cy.PublishtheApp();
-    cy.get(viewWidgetsPage.chartWidget).should("not.exist");
-  });
-
   afterEach(() => {
     cy.get(publish.backToEditor).click({ force: true });
   });
